Use --ignore-env in CI for better reproducibility
This makes sure that the CI can't use anything from the underlying
runner. In particular this doesn't provide the PATH, so we can't use
programs installed in the runner; everything we use must be defined in
the flake. I can't see any reason to _not_ do this so here we go.